fixing error, aturan pakai tipe data

+++ database/migrations/2020_12_14_122821_create_obat_resep_table.php
@@ -19,7 +19,7 @@ class CreateObatResepTable extends Migration
             $table->unsignedBigInteger('id_resep');
             $table->string('dosis');
             $table->string('aturan_pakai');
-            $table->integer('takaran_minum');
+            $table->float('takaran_minum');
             $table->enum('waktu_minum', ['Sebelum Makan', 'Saat Makan','Sesudah Makan']);
             $table->enum('keterangan', ['Kondisional', 'Harus Habis', 'Rutin']);
             $table->string('jml_obat');
